requirements.txt
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fuzzywuzzy==0.18.0
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
levenshtein==0.27.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
python-levenshtein==0.27.1
pytz==2025.1
pyzmq==26.2.1
rapidfuzz==3.12.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
xlrd==2.0.1
zipp==3.21.0
